+++ src/main/java/tools/InputTool.java
@@ -1,0 +1,4 @@
+package tools;
+
+public class InputTool {
+}
