+++ src/main/java/tools/PrintTool.java
@@ -1,0 +1,4 @@
+package tools;
+
+public class PrintTool {
+}
