+++ src/main/java/tools/Utility.java
@@ -1,0 +1,4 @@
+package tools;
+
+public class Utility {
+}
